requirements.txt
antlr4-python3-runtime==4.8
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-lru==2.0.4
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
charset-normalizer==3.1.0
comm==0.1.3
datasets==2.14.2
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
exceptiongroup==1.1.2
executing==1.2.0
fastjsonschema==2.18.0
fqdn==1.5.1
fsspec==2023.6.0
huggingface-hub==0.16.4
hydra-core==1.1.0
importlib-resources==6.0.0
ipdb==0.13.13
ipykernel==6.25.0
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
json5==0.9.14
jsonpointer==2.4
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
lightning-utilities==0.9.0
loggingext @ git+https://github.com/IGITUGraz/loggingext.git@906a5dc172a5620777b1886b06b4d1e06b8704bd
matplotlib-inline==0.1.6
mistune==3.0.1
multiprocess==0.70.14
nbclient==0.8.0
nbconvert==7.7.3
nbformat==5.9.2
nest-asyncio==1.5.7
notebook==7.0.1
notebook_shim==0.2.3
omegaconf==2.1.2
overrides==7.3.1
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pkgutil_resolve_name==1.3.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.1
Pygments==2.15.1
python-json-logger==2.0.7
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.0
regex==2023.6.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.9.2
safetensors==0.3.1
Send2Trash==1.8.2
SimManager @ git+https://github.com/IGITUGraz/SimManager.git@f823e66287b86f53d78ad3ca8940756bc66a16dd
sniffio==1.3.0
soupsieve==2.4.1
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torchmetrics==1.0.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
uri-template==1.3.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.8
xxhash==3.3.0
